+++ CHANGELOG.md
@@ -54,6 +54,7 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 * Properly sort map features with lifecycle prefixes in the _Past/Futures_ features ([#7582])
 * Only consider features with either `landuse`, `natural`, `amentiy` or `leisure` tag to be classified as _Landuse_ areas
 * Fix address field overwriting existing data when switching selected map features under certain circumstances ([#10260])
+* Show `name` field also when only a localized name is present (e.g. only `name:xy`, but not `name`) and the preset does not show the name field by default ([#10323], thanks [@samhoooo])
 #### :earth_asia: Localization
 #### :hourglass: Performance
 #### :mortar_board: Walkthrough / Help
@@ -72,7 +73,9 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 [#10257]: https://github.com/openstreetmap/iD/pull/10257
 [#10260]: https://github.com/openstreetmap/iD/issues/10260
 [#10302]: https://github.com/openstreetmap/iD/issues/10302
+[#10323]: https://github.com/openstreetmap/iD/issues/10323
 [@zbycz]: https://github.com/zbycz
+[@samhoooo]: https://github.com/samhoooo
 
 
 # 2.29.0